trying to figure out balance

See merge request campuses/cle/jan-2022/java/student-pairs/java-capstone-module-2-team-0!16

+++ capstone/tenmo-server/src/main/java/com/techelevator/tenmo/controller/AccountController.java
@@ -24,8 +24,8 @@ public class AccountController {
 
 
     @RequestMapping(path = "balance/{userid}", method = RequestMethod.GET)
-    public Integer getBalance(@Valid @PathVariable int userid) {
-        Integer balance = accountDao.getBalance(userid);
+    public BigDecimal getBalance(@Valid @PathVariable int userid) {
+        BigDecimal balance = accountDao.getBalance(userid);
 
         return balance;
     }
